fix: 修复AI决策摘要显示问题 + 前端缓存控制
- 新增结构化决策解析器 (structuredDecision.ts)
- 数据库添加 structured_decision 字段存储解析后的JSON
- 前端三层防缓存策略 (HTML meta + fetch cache + Hono中间件)
- 修复技术指标面板数据断点 (trading_signals表字段补全)
- 回填历史决策记录的结构化数据 (13条记录)

+++ src/database/close-and-reset.ts
@@ -115,7 +115,7 @@ async function closeAllPositions(): Promise<void> {
     logger.info("📊 获取当前持仓...");
     
     const positions = await exchangeClient.getPositions();
-    const activePositions = positions.filter((p: any) => Number.parseInt(p.size || "0") !== 0);
+    const activePositions = positions.filter((p: any) => Number.parseFloat(p.size || "0") !== 0);
     
     if (activePositions.length === 0) {
       logger.info("✅ 当前无持仓，跳过平仓");
@@ -125,7 +125,7 @@ async function closeAllPositions(): Promise<void> {
     logger.warn(`⚠️  发现 ${activePositions.length} 个持仓，开始平仓...`);
     
     for (const pos of activePositions) {
-      const size = Number.parseInt(pos.size || "0");
+      const size = Number.parseFloat(pos.size || "0");
       const contract = pos.contract;
       const symbol = contract.replace("_USDT", "");
       const side = size > 0 ? "多头" : "空头";
@@ -246,7 +246,7 @@ async function syncPositions(): Promise<void> {
     
     // 从交易所获取持仓
     const positions = await exchangeClient.getPositions();
-    const activePositions = positions.filter((p: any) => Number.parseInt(p.size || "0") !== 0);
+    const activePositions = positions.filter((p: any) => Number.parseFloat(p.size || "0") !== 0);
     
     logger.info(`📊 ${exchangeName} 当前持仓数: ${activePositions.length}`);
     
@@ -259,7 +259,7 @@ async function syncPositions(): Promise<void> {
       logger.info(`🔄 同步 ${activePositions.length} 个持仓到数据库...`);
       
       for (const pos of activePositions) {
-        const size = Number.parseInt(pos.size || "0");
+        const size = Number.parseFloat(pos.size || "0");
         if (size === 0) continue;
         
         const symbol = pos.contract.replace("_USDT", "");
