+++ src/scheduler/stopLossMonitor.ts
@@ -499,7 +499,7 @@ async function checkStopLoss() {
     
     // 1. 获取所有持仓
     const gatePositions = await exchangeClient.getPositions();
-    const activePositions = gatePositions.filter((p: any) => Number.parseInt(p.size || "0") !== 0);
+    const activePositions = gatePositions.filter((p: any) => Number.parseFloat(p.size || "0") !== 0);
     
     if (activePositions.length === 0) {
       // 清空内存记录
@@ -511,7 +511,7 @@ async function checkStopLoss() {
     
     // 2. 检查每个持仓
     for (const pos of activePositions) {
-      const size = Number.parseInt(pos.size || "0");
+      const size = Number.parseFloat(pos.size || "0");
       const symbol = pos.contract.replace("_USDT", "");
       const side = size > 0 ? "long" : "short";
       const quantity = Math.abs(size);
